+++ CHANGELOG.md
@@ -2,8 +2,35 @@
 
 All notable changes to the "synz" extension will be documented in this file.
 
+<!-- ## [Unreleased]
+
 ### Added
 
-- Auto-attaching (BETA)
+- Auto-attaching (BETA) -->
+
+## [Unreleased]
+
+### Added
+
+- Script browser - filter global, workspace, and synapse scripts.
+
+## [0.1.1] - 2024-09-08
+
+### Changes
+
+- Removed icon fill from the loader in the quick pick menu.
+- Removed details from the quick pick menu in unncessary places.
+- Disabled auto-attaching while fixes are made.
+- Updated icon for menu status bar button.
+
+### Fixes
+
+- Show executor buttons in quick pick menu when automaticallyDetectRobloxProcess is disabled
+
+## [0.1.0] - 2024-09-08
+
+### Added
+
 - Console logging
 - API for injecting/executing
+- Auto-attaching (BETA)